requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langchain_ollama
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
cassio
beautifulsoup4
langchain-groq
